requirements.txt
python-telegram-bot==22.6.0
psycopg[binary]==3.2.2
python-dotenv==1.0.0
